core-parse-opts: cast *ptr to int to avoid clang 3.4 warning
Clang warns when passing a char into isdigit(), so explicitly
promote it to an int to avoid the warning.

+++ core-parse-opts.c
@@ -92,7 +92,7 @@ static void ensure_numeric(const char *const str)
 	if (*ptr == '-')
 		ptr++;
 	while (*ptr) {
-		if (!isdigit(*ptr))
+		if (!isdigit((int)*ptr))
 			break;
 		ptr++;
 	}
